fix(internal): add/remove yank advice dynamically as needed
To avoid advising yank and yank-pop permanently on package load, we
instead keep track of how many buffers currently have yank-indent-mode
enabled, and add the required advices when the number goes above zero,
and remove them when it hits zero ago.

Annoyingly we have to advice kill-buffer to achieve an accurate
activate/deactivate count, as seems the kill-buffer-hook does not get
run for some buffers when they are killed.